requirements.txt
# Agentic Development Team - Dependencies
# Install with: pip install -r requirements.txt

# Core AutoGen packages
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0

# OpenAI client
openai>=1.0.0

# Async support
aiohttp>=3.8.0

# For MCP integration (optional, used in VS Code)
# mcp>=0.1.0

# YAML config parsing (for mcp_servers.yaml)
pyyaml>=6.0

# Pydantic for data validation
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Development tools (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
